Move some util functions
